moved ffmpeg logic to a custom hook
